M3 file

+++ src/main/java/sg/edu/nus/comp/cs4218/impl/app/MvApplication.java
@@ -17,28 +17,50 @@ import java.nio.file.*;
 
 public class MvApplication implements MvInterface {
 
+    /**
+     * default behaviour is override the existing file
+     */
     private boolean isOverride = true;
 
     @Override
-    public String mvSrcFileToDestFile(String srcFile, String destFile) throws Exception {
+    public String mvSrcFileToDestFile(String srcFile, String destFile) throws MvException {
         String destFilePath = FileSystemUtils.getAbsolutePathName(destFile);
         try {
+            // show check the file tree permit it be moved
+            if (FileSystemUtils.isSubDir(srcFile, destFile)) {
+                throw new MvException(srcFile + " is the sub dir of " + destFile + " or they are the same file.");
+            }
+            if (FileSystemUtils.isFileInFolder(srcFile, destFile)) {
+                throw new MvException("can not move, may source file is equal parent of dest file");
+            }
+            if (isOverride && new File(destFilePath).exists()) {
+                new File(destFile).delete();
+            }
             Files.move(Paths.get(FileSystemUtils.getAbsolutePathName(srcFile)),
                     Paths.get(destFilePath));
         } catch (NoSuchFileException e) {
             throw new MvException(ErrorConstants.ERR_FILE_NOT_FOUND + ":" + e.getMessage());
+        } catch (FileAlreadyExistsException e) {
+            throw new MvException("target file has existed:" + e.getMessage());
+
+        } catch (IOException e) {
+            throw new MvException(e.getMessage());
         }
         return destFilePath;
     }
 
 
     @Override
-    public String mvFilesToFolder(String destFolder, String... fileName) throws Exception {
+    public String mvFilesToFolder(String destFolder, String... fileName) throws MvException {
         String destFilePath = null;
         try {
             for (String oneFileName : fileName) {
                 destFilePath = FileSystemUtils.joinPath(FileSystemUtils.getAbsolutePathName(destFolder),
                         new File(FileSystemUtils.getAbsolutePathName(oneFileName)).getName());
+                // show check the file tree permit it be moved
+                if (FileSystemUtils.isFileInFolder(oneFileName, destFolder)) {
+                    throw new MvException("can not move, may source file is equal parent of dest file");
+                }
                 if (FileSystemUtils.isSubDir(oneFileName, destFolder)) {
                     throw new MvException(destFolder + " is the sub dir of " + destFolder + " or they are the same file.");
                 }
@@ -54,6 +76,10 @@ public class MvApplication implements MvInterface {
                     + "directory '" + destFolder + "' and cannot be replaced.").initCause(e);
         } catch (AccessDeniedException e) {
             throw new MvException(ErrorConstants.ERR_NO_PERM + ":" + e.getFile());
+        } catch (NoSuchFileException e) {
+            throw new MvException(ErrorConstants.ERR_FILE_NOT_FOUND + ":" + e.getMessage());
+        } catch (IOException e) {
+            throw new MvException(e.getMessage());
         }
         return destFilePath;
     }
@@ -71,14 +97,14 @@ public class MvApplication implements MvInterface {
                 throw new InvalidArgsException(ErrorConstants.ERR_MISSING_ARG);
             }
             String destPath = mvArgsParser.getDestFilePathName();
+            isOverride = mvArgsParser.isOverwrite();
             if (new File(FileSystemUtils.getAbsolutePathName(destPath)).isDirectory()) {
-                isOverride = mvArgsParser.isOverwrite();
                 mvFilesToFolder(destPath, toMoveFiles);
             } else {
                 if (toMoveFiles.length != 1) {
                     throw new InvalidArgsException(ErrorConstants.ERR_MISSING_ARG);
                 }
-                if (!mvArgsParser.isOverwrite() && new File(destPath).exists()) {
+                if (!isOverride && new File(destPath).exists()) {
                     throw new MvException("Destination file '" + destPath + "' already exists and cannot be replaced.");
                 }
                 mvSrcFileToDestFile(toMoveFiles[0], destPath);
@@ -92,6 +118,8 @@ public class MvApplication implements MvInterface {
             } catch (IOException ex) {
                 throw (MvException) new MvException("Could not write to output stream").initCause(ex);
             }
+        } finally {
+            isOverride = true;
         }
 
     }
